* Removed redundant test case on config-embedded security data source (covered by RPCSecurityManagerTest unit test) * Renamed UserAuthServiceTests as AuthDBTest, and moved to parent directory * Validate password encryption in the DB scenario
